MitMub/final-pjt-back/requirements.txt
asgiref==3.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.1.3
django-allauth==0.50.0
django-appconf==1.0.5
django-cors-headers==3.5.0
django-imagekit==4.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==3.4
oauthlib==3.2.2
pilkit==2.0
Pillow==9.3.0
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.12
